v2: 顾问端方案A重构 + 报告/客户档案修复
底栏改 陪伴/接诊(并入待整理)/●录音FAB/报告/客户;多主题换肤(暖玉柔光4套);
客户价值预测富渲染(对齐 web customer_profile)。

报告页:
- 修「整份报告解析炸→白页」根因:AI 报告字段类型会飘(timestamp_seconds 给小数、
  customer_tags 给字符串),Moshi 整份严格解析一处不符就全废。新增 SessionReportAdapter
  逐 PART 容错解析,单个坏字段降级为占位、不连累整份。
- PART01 全维度评估总览改手机适配竖排彩色标签块(顾客价值/痛点识别/销售诊断/质检评分),
  不再是挤窄的两列表;删「可操作建议」(改 prompt 生成)。
- 质检评分按子项均值重算(原直接读后端存量值显示 0)。
- 「查看详细评分」面板:右上角 ✕ 关闭 + 可纵向滚动(治"3.4 之后被裁切") + 子项进度条。

客户档案:
- 客户价值预测加「生成」按钮(放开为任何登录角色可点即生成)。
- 陪伴时间线每行可点 → 跳该次接诊分析报告。

绑定流程:接诊包/待整理合并;退回待整理(无理由);新客只姓名+尾号。

+++ android_app_v2/app/src/main/java/com/airec/bledemo/data/model/SessionReport.kt
@@ -185,7 +185,8 @@ data class CaseReview(
     @Json(name = "kind") val kind: String? = null,         // good | bad | miss
     @Json(name = "title") val title: String? = null,
     @Json(name = "quote") val quote: String? = null,
-    @Json(name = "timestamp_seconds") val timestampSeconds: Int? = null,
+    // AI 可能产出小数秒（如 2.81）；用 Double 兜底，避免 Moshi「Expected an int」整份报告解析失败
+    @Json(name = "timestamp_seconds") val timestampSeconds: Double? = null,
     @Json(name = "timestamp_label") val timestampLabel: String? = null,
     @Json(name = "segment") val segment: Int? = null,
     @Json(name = "surface") val surface: String? = null,   // 表层做法/问题
